pyproject.toml
[tool.poetry]
name = "habit"
version = "0.1.0"
description = "Habitat Analysis Tool for Medical Images"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.21.0"
scipy = "^1.7.0"
scikit-learn = "^1.0.0"
pandas = "^1.3.0"
matplotlib = "^3.4.0"
pyyaml = "^6.0"
SimpleITK = "^2.1.0"
pyradiomics = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.5b2"
mypy = "^0.910"
pylint = "^2.9.0"
pre-commit = "^2.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0903",  # too-few-public-methods
] 